data

+++ src/main/java/org/thyone/teamme/command/team/TeamDeleteCommand.java
@@ -3,9 +3,9 @@ package org.thyone.teamme.command.team;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.TextComponent;
 import net.kyori.adventure.text.format.NamedTextColor;
-import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
 import org.jetbrains.annotations.NotNull;
+import org.thyone.teamme.Protelum;
 import org.thyone.teamme.model.SubCommand;
 import org.thyone.teamme.model.SubCommandSyntax;
 import org.thyone.teamme.model.Team;
@@ -27,12 +27,14 @@ public class TeamDeleteCommand extends SubCommand {
 
     @Override
     public SubCommandSyntax[] getSyntax() {
-        return new SubCommandSyntax[0];
+        return new SubCommandSyntax[]{
+                new TeamDeleteNameSyntax()
+        };
     }
 
     @Override
     public TextComponent[] execute(Player player, String[] args) {
-        Team team = TeamStorage.getTeamOwn(player.getUniqueId().toString());
+        Team team = TeamStorage.getTeamOwn(player.getUniqueId());
         if (team == null)
             return new TextComponent[]{
                     Component
@@ -40,10 +42,17 @@ public class TeamDeleteCommand extends SubCommand {
                             .color(NamedTextColor.RED)
             };
 
+        if(!team.name.equals(args[0]))
+            return new TextComponent[]{
+                    Component
+                            .text("your givin name does not match with your team name")
+                            .color(NamedTextColor.RED)
+            };
+
         try {
             TeamStorage.delete(team.uuid);
         } catch (IOException exception) {
-            Bukkit.getLogger().log(Level.WARNING, exception.getMessage(), exception.getCause());
+            Protelum.getPlugin().getLogger().log(Level.WARNING, exception.getMessage(), exception.getCause());
 
             return new TextComponent[]{
                     Component
